Fix flax in time logic
I assumed that when 'Math.round(diff / 1000)' doesn't return 1, it will be
interpreted as false. This did however not happen as JavaScript seems to
interprete all numbers as "true". That's why I'm now checking against 1,
the number of seconds we actually wantI assumed that when 'Math.round(diff
/ 1000)' doesn't return 1, it will be interpreted as false. This did
however not happen as JavaScript seems to interprete all numbers as
"true". That's why I'm now checking against 1, the number of seconds we
actually want..