Add searchQuery to ng-show
This means that we always filter all servers with less than zero players,
unless a searchQuery is specific. This should be useful as most people
would probably expect all server matching a certain string to show up and
not only the ones with players. Especially if you are looking for, say the
connection details of a specific server